requirements.txt
aiohttp
python-dotenv
aiogram
mixpanel
pyyaml
asyncpg
logfire[aiohttp]
